+++ internal/profile/android.go
@@ -31,6 +31,7 @@ type (
 		Signature    string          `json:"signature,omitempty"`
 		SourceFile   string          `json:"source_file,omitempty"`
 		SourceLine   uint32          `json:"source_line,omitempty"`
+		SourceCol    uint32          `json:"-"`
 		InApp        *bool           `json:"in_app"`
 	}
 
@@ -69,6 +70,7 @@ func (m AndroidMethod) Frame() frame.Frame {
 		Function: methodName,
 		InApp:    &inApp,
 		Line:     m.SourceLine,
+		Column:   m.SourceCol,
 		MethodID: m.ID,
 		Package:  className,
 		Path:     m.SourceFile,
@@ -237,6 +239,10 @@ func (p *Android) FixSamplesTime() {
 
 // CallTrees generates call trees for a given profile.
 func (p Android) CallTrees() map[uint64][]*nodetree.Node {
+	return p.CallTreesWithMaxDepth(MaxStackDepth)
+}
+
+func (p Android) CallTreesWithMaxDepth(maxDepth int) map[uint64][]*nodetree.Node {
 	// in case wall-clock.secs is not monotonic, "fix" it
 	p.FixSamplesTime()
 
@@ -251,6 +257,7 @@ func (p Android) CallTrees() map[uint64][]*nodetree.Node {
 	buildTimestamp := p.TimestampGetter()
 	treesByThreadID := make(map[uint64][]*nodetree.Node)
 	stacks := make(map[uint64][]*nodetree.Node)
+	stackDepth := make(map[uint64]int)
 
 	methods := make(map[uint64]AndroidMethod)
 	for _, m := range p.Methods {
@@ -279,7 +286,6 @@ func (p Android) CallTrees() map[uint64][]*nodetree.Node {
 
 		switch e.Action {
 		case EnterAction:
-			enterPerMethod[e.MethodID]++
 			m, exists := methods[e.MethodID]
 			if !exists {
 				methods[e.MethodID] = AndroidMethod{
@@ -288,6 +294,11 @@ func (p Android) CallTrees() map[uint64][]*nodetree.Node {
 					Name:      "unknown",
 				}
 			}
+			stackDepth[e.ThreadID]++
+			if stackDepth[e.ThreadID] > maxDepth {
+				continue
+			}
+			enterPerMethod[e.MethodID]++
 			n := nodetree.NodeFromFrame(m.Frame(), ts, 0, 0)
 			if len(stacks[e.ThreadID]) == 0 {
 				treesByThreadID[e.ThreadID] = append(treesByThreadID[e.ThreadID], n)
@@ -298,6 +309,10 @@ func (p Android) CallTrees() map[uint64][]*nodetree.Node {
 			stacks[e.ThreadID] = append(stacks[e.ThreadID], n)
 			n.Fingerprint = generateFingerprint(stacks[e.ThreadID])
 		case ExitAction, UnwindAction:
+			stackDepth[e.ThreadID]--
+			if stackDepth[e.ThreadID] > maxDepth {
+				continue
+			}
 			if len(stacks[e.ThreadID]) == 0 {
 				continue
 			}
@@ -378,6 +393,10 @@ func (p *Android) NormalizeMethods(pi profileInterface) {
 }
 
 func (p Android) Speedscope() (speedscope.Output, error) {
+	return p.SpeedscopeWithMaxDepth(MaxStackDepth)
+}
+
+func (p Android) SpeedscopeWithMaxDepth(maxDepth int) (speedscope.Output, error) {
 	// in case wall-clock.secs is not monotonic, "fix" it
 	p.FixSamplesTime()
 
@@ -402,6 +421,7 @@ func (p Android) Speedscope() (speedscope.Output, error) {
 					Inline:        true,
 					IsApplication: inApp,
 					Line:          m.SourceLine,
+					Col:           m.SourceCol,
 					Name:          m.Name,
 				})
 			}
@@ -425,6 +445,7 @@ func (p Android) Speedscope() (speedscope.Output, error) {
 				Name:          fullMethodName,
 				File:          method.SourceFile,
 				Line:          method.SourceLine,
+				Col:           method.SourceCol,
 				IsApplication: inApp,
 				Image:         packageName,
 			})
@@ -458,6 +479,7 @@ func (p Android) Speedscope() (speedscope.Output, error) {
 
 	threadIDToProfile := make(map[uint64]*speedscope.EventedProfile)
 	methodStacks := make(map[uint64][]uint64) // map of thread ID -> stack of method IDs
+	stackDepth := make(map[uint64]int)
 	buildTimestamp := p.TimestampGetter()
 
 	enterPerMethod := make(map[uint64]int)
@@ -480,10 +502,18 @@ func (p Android) Speedscope() (speedscope.Output, error) {
 
 		switch event.Action {
 		case "Enter":
+			stackDepth[event.ThreadID]++
+			if stackDepth[event.ThreadID] > maxDepth {
+				continue
+			}
 			enterPerMethod[event.MethodID]++
 			methodStacks[event.ThreadID] = append(methodStacks[event.ThreadID], event.MethodID)
 			emitEvent(prof, speedscope.EventTypeOpenFrame, event.MethodID, ts)
 		case "Exit", "Unwind":
+			stackDepth[event.ThreadID]--
+			if stackDepth[event.ThreadID] > maxDepth {
+				continue
+			}
 			stack := methodStacks[event.ThreadID]
 			if len(stack) == 0 {
 				// This case happens when we filter events for a given transaction.