+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Label all node frames as system ([#392](https://github.com/getsentry/vroom/pull/392))
 - Fix react-native (android) rendering issue ([#397](https://github.com/getsentry/vroom/pull/397))
+- Restrict max depth for android profiles ([#424](https://github.com/getsentry/vroom/pull/424)), ([#426](https://github.com/getsentry/vroom/pull/426))
 
 **Internal**:
 
@@ -45,6 +46,7 @@
 - Bump trufflesecurity/trufflehog from 3.68.3 to 3.68.4 ([#419](https://github.com/getsentry/vroom/pull/419))
 - Bump trufflesecurity/trufflehog from 3.68.4 to 3.68.5 ([#420](https://github.com/getsentry/vroom/pull/420))
 - Bump trufflesecurity/trufflehog from 3.68.5 to 3.69.0 ([#421](https://github.com/getsentry/vroom/pull/421))
+- Keep col info when converting from android method to frame ([#423](https://github.com/getsentry/vroom/pull/423))
 
 ## 23.12.0